requirements.txt
streamlit==1.28.0
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
plotly==5.15.0
seaborn==0.12.2
matplotlib==3.7.2
mysql-connector-python==8.1.0